+++ backend/canisters/group_index/impl/src/updates/mod.rs
@@ -8,6 +8,7 @@ pub mod c2c_make_community_private;
 pub mod c2c_make_private;
 pub mod c2c_mark_active;
 pub mod c2c_mark_community_active;
+pub mod c2c_start_importing_group_into_community;
 pub mod c2c_update_community;
 pub mod c2c_update_group;
 pub mod delete_frozen_group;